% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg2shiny.R
\name{gg2shiny}
\alias{gg2shiny}
\title{Converts ggplot2 plots into a shiny application}
\usage{
gg2shiny(ggplotobj)
}
\arguments{
\item{ggplotobj}{the plot to be exported}
}
\description{
Converts ggplot2 plots into a shiny application
}
\examples{
\dontrun{
library(gg2shiny)
library(ggplot2)
pl <- ggplot(iris, aes(Sepal.Length, Petal.Length))+geom_point()
gg2shiny(pl)
}
}

